ops:make `toolchain.sh` smarter

+++ lumium-space/scripts/toolchain.sh
@@ -20,9 +20,11 @@ if which clang >/dev/null; then
     echo "clang toolchain already installed"
 else
     echo "installing clang toolchain"
-    rm -rf clang+llvm-14.0.0-x86_64-linux-gnu-ubuntu-18.04
-    curl https://github.com/llvm/llvm-project/releases/download/llvmorg-14.0.0/clang+llvm-14.0.0-x86_64-linux-gnu-ubuntu-18.04.tar.xz -L --output clang.tar.xz
-    tar -xvf clang.tar.xz 
-    rm -rf clang.tar.xz
+    if [ ! -d "clang+llvm-14.0.0-x86_64-linux-gnu-ubuntu-18.04" ]; then 
+        rm -rf clang+llvm-14.0.0-x86_64-linux-gnu-ubuntu-18.04
+        curl https://github.com/llvm/llvm-project/releases/download/llvmorg-14.0.0/clang+llvm-14.0.0-x86_64-linux-gnu-ubuntu-18.04.tar.xz -L --output clang.tar.xz
+        tar -xvf clang.tar.xz
+        rm -rf clang.tar.xz
+    fi
     export PATH="$PATH:$PWD/clang+llvm-14.0.0-x86_64-linux-gnu-ubuntu-18.04/bin"
 fi
